+++ packages/webapp/src/store/folders-store.ts
@@ -10,6 +10,12 @@ export interface FoldersStore {
    */
   view: TFolderView | ''
   setView: (view: TFolderView) => void
+  /**
+   * Thumbnail size of the user as step around the default size. It is empty as
+   * long as the user did not change the size, than the configured size is used
+   */
+  sizeStep: number | ''
+  setSizeStep: (sizeStep: number) => void
 }
 
 export const useFoldersStore = create<
@@ -23,6 +29,12 @@ export const useFoldersStore = create<
 
     setView: (view: TFolderView) => {
       set((state) => ({...state, view}))
+    },
+
+    sizeStep: '',
+
+    setSizeStep: (sizeStep: number) => {
+      set((state) => ({...state, sizeStep}))
     },
   }), { name: 'gallery-folders' })
 )